eigth 42

+++ 42Cursus/Libft/ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: ellaca-f <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 15:48:45 by ellaca-f          #+#    #+#             */
-/*   Updated: 2019/11/07 12:54:08 by ellaca-f         ###   ########.fr       */
+/*   Updated: 2019/11/14 20:58:20 by ellaca-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 void	*ft_calloc(size_t count, size_t size)
 {
 	char	*s;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	s = malloc(count * size);
